Add the ability to return UNIX timestamps, and approximate timestamps

+++ src/helpers.rs
@@ -9,11 +9,11 @@ pub fn _millis_to_u64(millis: u64) -> u64 {
 }
 
 #[inline]
-pub fn _timespec_to_u64(tp_sec: u64, tp_nsec: u64) -> u64 {
-    (tp_sec << 32) | ((tp_nsec * 9_223_372_037) >> 31)
+pub fn _timespec_to_u64(tp_sec: u64, tp_nsec: u32) -> u64 {
+    (tp_sec << 32) | ((tp_nsec as u64 * 9_223_372_037) >> 31)
 }
 
 #[inline]
-pub fn _timeval_to_u64(tv_sec: u64, tv_usec: u64) -> u64 {
-    (tv_sec << 32) | ((tv_usec * 9_223_372_036_855) >> 31)
+pub fn _timeval_to_u64(tv_sec: u64, tv_usec: u32) -> u64 {
+    (tv_sec << 32) | ((tv_usec as u64 * 9_223_372_036_855) >> 31)
 }
